requirements.txt
firebase-admin==6.5.0
Flask==3.0.3
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
python-dotenv==1.0.1
scikit-learn==1.4.2
